+++ tests/unit/variable-list-option.spec.ts
@@ -63,9 +63,9 @@ describe('Variable List option', () => {
       });
       await wrapper.vm.$nextTick();
     });
-    it('should display date value in UTC timezone', () => {
+    it('should display date value with default toString', () => {
       expect(wrapper.find('.widget-variable-option__value').text()).toBe(
-        'Fri, 11 Jun 2021 08:09:17 GMT',
+        new Date(1623398957013).toString(),
       );
     });
   });
@@ -104,7 +104,7 @@ describe('Variable List option', () => {
       {
         type: 'date',
         value: new Date(1623398957013),
-        attendedValue: `"${new Date(1623398957013).toUTCString()}"`,
+        attendedValue: `"${new Date(1623398957013).toString()}"`,
       },
     ].forEach(
       ({ type, value, attendedValue }: { type: string; value: any; attendedValue: any }) => {